add root/bitarray.d
