+++ src/dmd/arraytypes.h
@@ -11,6 +11,7 @@
 #pragma once
 
 #include "root/array.h"
+#include "root/bitarray.h"
 
 typedef Array<class TemplateParameter *> TemplateParameters;
