+++ src/dmd/compiler.h
@@ -11,6 +11,7 @@
 #pragma once
 
 #include "root/array.h"
+#include "root/bitarray.h"
 
 // This file contains a data structure that describes a back-end compiler
 // and implements compiler-specific actions.
