+++ src/dmd/root/array.h
@@ -206,21 +206,3 @@ struct Array
     }
 };
 
-struct BitArray
-{
-    BitArray()
-      : len(0)
-      , ptr(NULL)
-    {}
-
-    ~BitArray()
-    {
-        mem.xfree(ptr);
-    }
-
-    d_size_t len;
-    d_size_t *ptr;
-
-private:
-    BitArray(const BitArray&);
-};
